+++ publish_queue.py
@@ -1,7 +1,7 @@
 from pyxavi.logger import Logger
 from pyxavi.config import Config
-from src.mastodon_helper import MastodonHelper
-from src.publisher import Publisher
+from publisher import Publisher
+from definitions import ROOT_DIR
 
 class QueuePublisher:
     '''
@@ -10,6 +10,9 @@ class QueuePublisher:
     def init(self):
         self._config = Config()
         self._logger = Logger(self._config).get_logger()
+        self._publisher = Publisher(
+            config=self._config, base_path=ROOT_DIR
+        )
 
         return self
 
@@ -18,14 +21,8 @@ class QueuePublisher:
         Just publishes the queue
         '''
         try:
-            # All actions are done under a Mastodon API instance
-            mastodon = MastodonHelper.get_instance(self._config)
-
-            # Read from the queue the toots to publish
-            # and publishes all of them
-            publisher = Publisher(self._config, mastodon)
             self._logger.info("Publishing the whole queue")
-            publisher.publish_all_from_queue()
+            self._publisher.publish_all_from_queue()
         except Exception as e:
             self._logger.exception(e)